+++ opensearch/src/main/java/org/opensearch/sql/opensearch/client/OpenSearchNodeClient.java
@@ -180,7 +180,11 @@ public class OpenSearchNodeClient implements OpenSearchClient {
             }
           });
     } else if (request instanceof OpenSearchQueryRequest) {
-      request.clean(pitId -> {});
+      request.clean(
+          pitId -> {
+            DeletePitRequest deletePitRequest = new DeletePitRequest(pitId);
+            deletePit(deletePitRequest);
+          });
     }
   }
